schedule styling: day-dive resized and given equal spacing, component title altered, back button included, and scrollabillity implemented
