+++ server/public/styles.css
@@ -179,15 +179,8 @@ height:75px;
   padding:10px;
 }
 
-.detailsInfo{
-/* font-family: 'Oswald', sans-serif; */
-}
-.reviewsInfo{
-/* font-family: 'Oswald', sans-serif; */
-}
-
 .eventList{
-  height:65%;
+  /* height:65%; */
   width:90%;
   background: white;
   border: 4px solid #000000;
@@ -199,6 +192,7 @@ height:75px;
   top: 15%;
   margin-left: 5%;
   box-shadow: 7px 7px 7px rgba(0, 0, 0, 0.25);
+    text-shadow: 0px 3px 5px rgba(0, 0, 0, 0.25);
   border-radius: 10px;
   padding: 3%
 }
@@ -307,7 +301,7 @@ height:75px;
   border: none;
 }
 
-.settingsContainer {
-  border-radius: 10px;
-  text-shadow: 0px 3px 5px rgba(0, 0, 0, 0.25);
+.selected-day-schedule {
+  background-color: #2a9d8f;
+  color: white;
 }
